frontend-*: initial refactor to create and use -internal
