+++ packages/frontend-internal/README.md
@@ -1,0 +1,3 @@
+# @internal/frontend
+
+This is an internal package used by the other frontend packages. It does not get published to NPM, but instead inlined into consuming packages due to the `backstage.inline` flag in `package.json`.
