+++ packages/frontend-internal/catalog-info.yaml
@@ -1,0 +1,9 @@
+apiVersion: backstage.io/v1alpha1
+kind: Component
+metadata:
+  name: internal-frontend
+  title: '@internal/frontend'
+spec:
+  lifecycle: experimental
+  type: backstage-web-library
+  owner: maintainers
